refactor namespace for custom views
