+++ app/src/main/res/layout/home_toolbar.xml
@@ -6,7 +6,7 @@
     android:layout_height="?attr/actionBarSize"
     android:background="@color/colorPrimary">
 
-    <za.co.app.budgetbee.ui.views.MonthSwitcher
+    <za.co.app.budgetbee.ui.custom_views.MonthSwitcher
         android:id="@+id/month_switcher"
         android:layout_width="wrap_content"
         android:layout_height="@dimen/year_switcher_height"
